+++ ControleBiblioteca.csproj
@@ -11,6 +11,12 @@
   </ItemGroup>
 
   <ItemGroup>
+    <Compile Update="FormAluguel.cs">
+      <SubType>Form</SubType>
+    </Compile>
+    <Compile Update="FormLivro.cs">
+      <SubType>Form</SubType>
+    </Compile>
     <Compile Update="Properties\Resources.Designer.cs">
       <DesignTime>True</DesignTime>
       <AutoGen>True</AutoGen>
