+++ FormAluguel.cs
@@ -1,137 +1,64 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.ComponentModel;
 using System.Data;
+using System.Drawing;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using MySql.Data;
-using MySql.Data.MySqlClient;
+using System.Windows.Forms;
 
 namespace ControleBiblioteca {
-    class FormAluguel {
-        public int idaluguel { get; set; }
-        public int idpessoa { get; set; }
-        public int idendereco { get; set; }
-        public DateTime datasaida { get; set; }
-        public DateTime dataretorno { get; set; }
-        public double valor { get; set; }
+    public partial class FormAluguel : Form {
+
+        int indexSelecionadoAluguel;
 
         public FormAluguel() {
+            InitializeComponent();
+        }
+
+        public void buscaAlugueis() {
 
         }
 
-        public FormAluguel(int idpessoa, int idendereco, DateTime datasaida, DateTime dataretorno, double valor) {
-            this.idpessoa = idpessoa;
-            this.idendereco = idendereco;
-            this.datasaida = datasaida;
-            this.dataretorno = dataretorno;
-            this.valor = valor;
+        private void btnNovoAluguel_Click(object sender, EventArgs e) {
+            indexSelecionadoAluguel = -1;
+            btnSalvarAluguel.Text = "Cadastrar";
+            cmbPessoa.Text = string.Empty;
+            cmbLivroAlugado.Text = string.Empty;
+            txtDataSaida.Value = DateTime.Now;
+            txtDataRetorno.Value = DateTime.Now;
+            txtValor.Text = string.Empty;
+
+            tabControl1.SelectedTab = tabControl1.TabPages["tabAluguel"];
+            cmbPessoa.Focus();
         }
 
-        public FormAluguel(int idpessoa, int idendereco, DateTime datasaida, DateTime dataretorno, double valor, 
-            int idaluguel) {
-            this.idpessoa = idpessoa;
-            this.idendereco = idendereco;
-            this.datasaida = datasaida;
-            this.dataretorno = dataretorno;
-            this.valor = valor;
-            this.idaluguel = idaluguel;
-        }
-
-
-        public string insert() {
-            FormConexao c = new FormConexao();
-            try {
-                c.abreConexao();
-
-                MySqlCommand cmd = new MySqlCommand(
-                    @"", c.conexaoBD());
-
-                cmd.Parameters.AddWithValue("@idpessoa", this.idpessoa);
-                cmd.Parameters.AddWithValue("@idendereco", this.idendereco);
-                cmd.Parameters.AddWithValue("@datasaida", this.datasaida);
-                cmd.Parameters.AddWithValue("@dataretorno", this.dataretorno);
-                cmd.Parameters.AddWithValue("@valor", this.valor);
-                cmd.ExecuteNonQuery();
-                return "Ok";
+        private void btnSalvarAluguel_Click(object sender, EventArgs e) {
+            if (cmbPessoa.Text.Trim().Equals(string.Empty)) {
+                MessageBox.Show("Selecione uma pessoa!");
+                cmbPessoa.Focus();
+            } else if (cmbLivroAlugado.Text.Trim().Equals(string.Empty)) {
+                MessageBox.Show("Selecione um livro!");
+                cmbLivroAlugado.Focus();
+            } else if (txtValor.Text.Trim().Equals(string.Empty)) {
+                MessageBox.Show("Insira um valor!");
+                txtValor.Focus();
+            } else if (txtDataRetorno.Value < txtDataSaida.Value) {
+                MessageBox.Show("Data de retorno não pode ser menor do que entrada!");
             }
-            catch (Exception erro) {
-                return erro.Message;
-            }
-            finally {
-                c.fechaConexao();
-            }
-        }
-
-        public DataTable select() {
-            DataTable tabela = new DataTable();
-            FormConexao conexao = new FormConexao();
-            try {
-                conexao.abreConexao();
-
-                MySqlCommand cmd = new MySqlCommand("SELECT * FROM cad_alugueis", conexao.conexaoBD());
-                MySqlDataReader dados = cmd.ExecuteReader();
-                if (dados.HasRows == true) {
-                    tabela.Load(dados);
+            else {
+                Aluguel a = new Aluguel();
+                if (indexSelecionadoAluguel == -1) {
+                    a.insert();
                 }
                 else {
-                    tabela = new DataTable("Erro");
-                    tabela.Rows.Add("Nenhuma movimentação encontrada");
+                    a.idaluguel = indexSelecionadoAluguel;
+                    a.update();
                 }
-                return tabela;
-            }
-            catch (Exception erro) {
-                tabela = new DataTable("Erro");
-                tabela.Rows.Add("Erro: " + erro.Message);
-                return tabela;
-            }
-            finally {
-                conexao.fechaConexao();
-            }
-        }
 
-
-        public string update() {
-
-            FormConexao c = new FormConexao();
-            try {
-                c.abreConexao();
-                MySqlCommand cmd = new MySqlCommand(@"", c.conexaoBD());
-                cmd.Parameters.AddWithValue("@idpessoa", this.idpessoa);
-                cmd.Parameters.AddWithValue("@idendereco", this.idendereco);
-                cmd.Parameters.AddWithValue("@datasaida", this.datasaida);
-                cmd.Parameters.AddWithValue("@dataretorno", this.dataretorno);
-                cmd.Parameters.AddWithValue("@valor", this.valor);
-                cmd.Parameters.AddWithValue("@idaluguel", this.idaluguel);
-                cmd.ExecuteNonQuery();
-                return "Movimentação atualizada com sucesso!";
-            }
-            catch (Exception erro) {
-                return "Erro: " + erro.Message;
-            }
-            finally {
-                c.fechaConexao(); ;
-            }
-
-
-        }
-
-        public string delete(int codigo) {
-
-            FormConexao c = new FormConexao();
-            try {
-                c.abreConexao();
-                MySqlCommand cmd = new MySqlCommand(@"", c.conexaoBD());
-                cmd.Parameters.AddWithValue("@idaluguel", idaluguel);
-                cmd.ExecuteNonQuery();
-                return "Movimentação excluida com sucesso!";
-
-            }
-            catch (Exception erro) {
-                return "Erro: " + erro.Message;
-            }
-            finally {
-                c.fechaConexao();
+                buscaAlugueis();
+                tabControl1.SelectedTab = tabControl1.TabPages["tabAlugueisCadastrados"];
             }
         }
     }
