+++ Program.cs
@@ -14,7 +14,7 @@ namespace ControleBiblioteca {
             Application.SetHighDpiMode(HighDpiMode.SystemAware);
             Application.EnableVisualStyles();
             Application.SetCompatibleTextRenderingDefault(false);
-            Application.Run(new FormControle());
+            Application.Run(new FormPessoa());
         }
     }
 }
